Robustness: 64-bit arc ids, release-mode rc>=0 check, loud balance-guard warnings
Three guards against silent failure at scale, motivated by advertising
NISAR readiness (single-look frames are 60,000 x 60,000 = 3.6 Gpixel):

1. Widen arc ids to 64 bits in every predecessor store. Residual arc
   ids run to 2*num_forward (~16 per pixel), which overflows i32 past
   ~268 Mpixel and u32 past ~536 Mpixel - far below a single-look NISAR
   frame - and a wrapped id corrupts the pred chain (and the unwrap)
   with no error. ShortestPaths::pred_arc and both ssp.rs local pred
   arrays go i32 -> i64; the parallel Dial's Proposal tuple goes
   u32 -> u64. Node ids stay u32 in the hot bucket queues (2^32 nodes =
   ~65,535^2, so the 60k x 60k target fits) and are now guarded by an
   assert in RectangularGridGraph::new that fails loudly at
   construction instead of wrapping in the queues.

2. Check the non-negative reduced-cost invariant in RELEASE builds. The
   single-source Dial SSP relied on a debug_assert - compiled out
   exactly where it matters - and a negative rc silently poisons
   distances (entries land in buckets the scan already passed and are
   discarded as stale). Now a negative rc abandons that source's search
   with the network untouched, counts it as stranded for the BFS
   balance guard to pair, and warns unconditionally at the end: the
   output stays usable (bounded local error) and the invariant break is
   visible instead of silent.

3. Make the BFS balance guard loud on its own failure. Its stranded
   branch only printed under WHIRLWIND_DEBUG; if the guard of last
   resort cannot pair a residue the unwrap WILL carry a full-width 2pi
   tear, and silence there would recreate the original Ridgecrest bug
   with no trace in the logs. Stranded sources and any remaining excess
   now warn unconditionally.

Tests: grid guard (should_panic at 70,000^2; 60,001^2 accepted and
genuinely needs 64-bit arc ids), negative-rc recovery (corrupted
potential -> search abandoned untouched -> BFS guard pairs, balanced).
All 115 core tests pass. Ridgecrest crop repro bit-identical (97.42%,
same ambiguity histogram) at unchanged runtime.

+++ crates/whirlwind-core/src/shortest_path/dial.rs
@@ -161,7 +161,7 @@ pub fn run_into<G: ResidualGraph>(g: &G, net: &Network, sp: &mut ShortestPaths) 
             let nd = cur_dist + rc;
             if nd < sp.dist[v] {
                 sp.dist[v] = nd;
-                sp.pred_arc[v] = arc as i32;
+                sp.pred_arc[v] = arc as i64;
                 let b = (nd as usize) % k;
                 buckets[b].push(v as u32);
                 *pending += 1;
@@ -288,7 +288,7 @@ pub fn run_full_into<G: ResidualGraph>(g: &G, net: &Network, sp: &mut ShortestPa
             let nd = cur_dist + rc;
             if nd < sp.dist[v] {
                 sp.dist[v] = nd;
-                sp.pred_arc[v] = arc as i32;
+                sp.pred_arc[v] = arc as i64;
                 let b = (nd as usize) % k;
                 buckets[b].push_back(v as u32); // FIFO (see decl)
                 *pending += 1;
@@ -320,7 +320,10 @@ const PAR_THRESHOLD: usize = 256;
 /// Proposal emitted by a phase-1 thread: `(v, nd, arc)`.
 /// We collect into per-thread Vecs and reduce, so the constant matters less
 /// than amortized.
-type Proposal = (usize, i64, u32);
+type Proposal = (usize, i64, u64);
+// The arc id is u64, not u32: residual arc ids run to 2*num_forward (~16 x
+// pixels), which overflows u32 past ~536 Mpixel (NISAR single-look posting
+// is 3.6 Gpixel). Node ids stay u32 - guarded by the assert in grid.rs.
 
 /// Parallel Dial's multi-source Dijkstra. Functionally equivalent to [`run`],
 /// but each large-enough bucket is relaxed via rayon. See module docs for the
@@ -409,7 +412,7 @@ pub fn run_parallel<G: ResidualGraph>(g: &G, net: &Network) -> ShortestPaths {
                         let nd = cur_dist + rc;
                         if nd < sp.dist[v] {
                             sp.dist[v] = nd;
-                            sp.pred_arc[v] = arc as i32;
+                            sp.pred_arc[v] = arc as i64;
                             let b = (nd as usize) % k;
                             buckets[b].push(v as u32);
                             *pending += 1;
@@ -486,7 +489,7 @@ pub fn run_parallel<G: ResidualGraph>(g: &G, net: &Network) -> ShortestPaths {
                             };
                             let nd = cur_dist + rc;
                             if nd < sp_dist_snap[v] {
-                                props.push((v, nd, arc as u32));
+                                props.push((v, nd, arc as u64));
                             }
                         };
                         out_buf.clear();
@@ -522,7 +525,7 @@ pub fn run_parallel<G: ResidualGraph>(g: &G, net: &Network) -> ShortestPaths {
             for (v, nd, arc) in proposals {
                 if nd < sp.dist[v] {
                     sp.dist[v] = nd;
-                    sp.pred_arc[v] = arc as i32;
+                    sp.pred_arc[v] = arc as i64;
                     let b = (nd as usize) % k;
                     buckets[b].push(v as u32);
                     pending += 1;
